Debug - dtype, decive

+++ HW2/ddpg_cheetah.py
@@ -164,13 +164,13 @@ class DDPG(object):
         self.num_inputs = num_inputs
         self.action_space = action_space
 
-        self.actor = Actor(hidden_size, self.num_inputs, self.action_space).to(device)
-        self.actor_target = Actor(hidden_size, self.num_inputs, self.action_space).to(device)
-        self.actor_perturbed = Actor(hidden_size, self.num_inputs, self.action_space).to(device)
+        self.actor = Actor(hidden_size, self.num_inputs, self.action_space).to(device).float()
+        self.actor_target = Actor(hidden_size, self.num_inputs, self.action_space).to(device).float()
+        self.actor_perturbed = Actor(hidden_size, self.num_inputs, self.action_space).to(device).float()
         self.actor_optim = Adam(self.actor.parameters(), lr=lr_a)
 
-        self.critic = Critic(hidden_size, self.num_inputs, self.action_space).to(device)
-        self.critic_target = Critic(hidden_size, self.num_inputs, self.action_space).to(device)
+        self.critic = Critic(hidden_size, self.num_inputs, self.action_space).to(device).float()
+        self.critic_target = Critic(hidden_size, self.num_inputs, self.action_space).to(device).float()
         self.critic_optim = Adam(self.critic.parameters(), lr=lr_c, weight_decay=1e-2)
 
         self.gamma = gamma
